@@ -10,7 +10,6 @@ classifiers = [
     "License :: OSI Approved :: MIT License",
     "Programming Language :: Python :: 3",
     "Programming Language :: Python :: 3 :: Only",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
@@ -24,7 +23,7 @@ classifiers = [
     "Topic :: Software Development :: Libraries :: Python Modules",
 ]
 description = "Self-contained wheels for Windows"
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 dependencies = ["pefile >= 2024.8.26"]
 
 [project.scripts]
